Refactor source character set related test code
The one issue with this change is that now there are compiler warnings
complaining about dead code. The thing is, such code is being used, but
the fact that each tests file are treated like separate crates forces
you to declare the tests utility module in each one of them, thus any
code that is not used by such crates are warned about.

In other words, if some given function is used at X_test.rs, but not in
Y_test.rs, then the compiler will generate a dead code warning for such
function not being used in the Y_test.rs crate.