chore: move unit tests into a module
This is more in line with what I had in mind what would be the ideal way
of doing unit tests. It still needs you to declare the test module in
tests/mod.rs, but at least you can't import private stuff, and it's all
considered to be the same crate, thus generating correct coverage
statistics.