feat: translate line-feed + carriage-return to line-feed only
This patch reworks most of the internals of `CharStream` so that such
translation is easier to implement. Now, the stream advancing is done
through the peek function, the result of which is cached so that next
calls return the previously worked char. The consume function only
advances the byte position of the peeked char, and clears the cached
peek, so that next calls to peek advance to the next character in the
stream.